CC-410: Add syslogs for syslog-ng cert failures
According to the Common Criteria requirements, certificate validation
failures should be logged to syslog.  This adds the approriate syslog
calls to syslog-ng.

It also logs a warning if the OCSP connection could not be made.
This event doesn't cause certificate validation to fail, and so
is not an error, but it is worth a warning.

Change-Id: I3454992a8809cc266c57f30d6c4259227317022d
Reviewed-by: Ken Kato <ken_kato@alliedtelesis.com>
Reviewed-by: David Thomson <david.thomson@alliedtelesis.co.nz>
Reviewed-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>
Signed-off-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>