BLDSYS-198: Don't attempt to get local FQDN hostname from DNS
On a chassis, domain-name like hostnames are written to /etc/hosts
(ie awplus-1.5, with an offending '.'). The behavior of gethostbyname
is to check /etc/hosts before resolving with DNS.

The result of this is that messages will be logged on chassis with
the full stack member hostname, regardless of DNS configuration.

As such, while we write illegal hostnames to /etc/hosts for stacking
we cant support logging messages with FQDN.

Change-Id: I7dbab8dbbdd1f05559cabc6e25112e25ffbb8e0b
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>