CR-57961: Mark messages as consumed in a write failure.
Fix a crash that occurs following a write failure when the file writer
buffer becomes full.

If a write failure occurs in log_progo_file_writer_post as the buffer
becomes full, a subsequent call will be made where msg and
self->buffer[self->buff_count].iov_base point to the same memory.

The first call to log_proto_file_writer_flush will write the message
and free the buffer. The following call will raise an abort signal
when it attempts to free the buffer a second time.

If the message has been stored in self->buffer, it should be marked
as consumed irrespective of the result of log_proto_file_writer_flush.

Change-Id: Iaaa301f1706e2ba58366cd59152ae73507069a7e
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>
Signed-off-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>