CR-57983: Cherry-pick: affile: Add possibility to only signal re-open of file...
Cherry-pick upstream commits: (f38fa0f bugfixes acaee7c)

  (acaee7c) affile: Add possibility to only signal re-open of file handles

    Avoid reloading the whole configuration by having a designated signal
    (USR1) to re-open all file destination log writers' file handles.
    This avoids the problem of issue #1423 where applications like newsyslog
    can use this new signal after it has rotated the log files.

  (f38fa0f) affile: Do not iterate writer_hash if uninitialized

    Both single_writer and writer_hash can be uninitialized as the opening
    of a file is deferred until there is some message to write.

This allows the use of logrotate without reloading config
significantly reducing the impact of reload-triggered memory leaks.

Note. syslog-ng v3.12 has a number of memory leak fixes, but I would prefer
not to upgrade again this close to a release to avoid unnecessary risk.

It is believed that logrotate triggered reloads are the primary cause of
the memory leaks we've seen, and that removing them should be sufficient
to reduce the rate of memory loss below a noticable level.

Change-Id: Icd6d73c71b6952eb3817f41b3b6a37c1fde79ea2
Reviewed-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>