CR-58445: Remove invalid assertion from afsocket_dd_restore_writer.
Asserting that self->writer is null is an invalid assertion.

If the config being initialized is new, self->writer is guaranteed
to be null, however if we are restoring an old config after initializing
a new configuration failed, an existing writer may be present, if
keep-alive is off (Using keepalive keep-alive means writer is set to
null when it is stored as a persist item).

This mirrors the reload behavior of program destinations, which also
impliment a keepalive option.

Change-Id: Ib60b81c9bfae158b296c61cc0bd48da2a9a01e75
Reviewed-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>
Reviewed-by: Scott Parlane <scott.parlane@alliedtelesis.co.nz>
Signed-off-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>