CR-58712: Make sure the suspend timer is unregistered in logwriter deinit
log_writer_stop_watches is currently responsible for unregistering the
suspend timer when de-initializing a logwriter, though it only does this
if watches are enabled. In log_writer_work_finished() the suspend timer is
armed following a call to log_writer_broken(), which disables watches.

Make sure the suspend timer is unregistered regardless of whether watches
are enabled.

Change-Id: I7365e5b044ac266534e401fc28444e7839b8ac56
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>