CR-59869: Increment LogPipe ref counter when passed to TLSSession
When a TLS session is initiated for secure logging, the afsocket
module passes a reference to the AFInetDestDriver instance to the
TLSSession structure to use as verification data. Due to the
timing of configuration changes, the driver structure may be
freed before the TLSSession structure, meaning that TLSSession
will contain a dangling pointer to the AFInetDestDriver. The
afsocket module may try to verify the session using the invalid
data, which may result in a crash.

This patch adds two extra function pointers to the
TransportMapperInet structure, one that is used to increment the
driver's reference counter when it is passed to the TLSSession,
and one that is called to decrement the driver's reference
counter when the session structure is destroyed. This ensures
that the driver is only freed once there are no sessions
referencing it.

Change-Id: Ia07b6169c9579d1a2fcdbc4e608b94b030e6e072
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>