CR-62976: logwriter: Avoid null pointer access from suspended writer
The core dump shows that when IMISH exits while a terminal monitor is
enabled, syslog-ng causes SEGV by accessing a NULL pointer. It hasn't been
reproduced but it indicates that logging out from a telnet session
removes the tty ("/dev/pts/0") and syslog-ng fails to write.
Syslog-ng then intends to retry later and put the writer in a timer.
When the timer expires, it finds the writer client was already freed.
And accessing a NULL pointer causes SEGV.

It seems it's a rare timing issue regarding config reloading but could
not figure out exactly how the situation can happen.

So make sure the client is valid in log_writer_error_suspend_elapsed()
before accessing it (self->proto).

Change-Id: Iba44841c8e5e6bc21fe6be06b15d0035b3449c33
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>
Signed-off-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>