CC-318 Add OCSP validation to syslog-ng
This adds support in syslog-ng to use the OCSP data in the certificate
to check for revocation.  If the OCSP server can be contacted and
replies that the certificate has been revoked then the connection is
not made.

It also adds a check for the presence of the Extended Key Usage field.
The openssl client code already validates that the data in the field
is correct if it is present but the requirement (see CC-291) is that
the field must be present.

Change-Id: I1bad990685eec07b5f0c9dbfbf01e94d760cda36
Reviewed-by: Ken Kato <ken_kato@alliedtelesis.com>
Reviewed-by: John Culbreth <john_culbreth@alliedtelesis.com>
Reviewed-by: Chris Packham <chris.packham@alliedtelesis.co.nz>
Reviewed-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>
Reviewed-by: David Thomson <david.thomson@alliedtelesis.co.nz>
Signed-off-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>