VPNAUTH-29: Update previous patch for OpenSSL v1.1.1
* Change manual search for extended key usage to use a function call.
* Use accessor functions where required.

CC-318 Add OCSP validation to syslog-ng

This adds support in syslog-ng to use the OCSP data in the certificate
to check for revocation.  If the OCSP server can be contacted and
replies that the certificate has been revoked then the connection is
not made.

It also adds a check for the presence of the Extended Key Usage field.
The openssl client code already validates that the data in the field
is correct if it is present but the requirement (see CC-291) is that
the field must be present.

Change-Id: I6b14fcb46ff20b298ad646ffe74b1cfd82b6ff6e
Reviewed-by: Anthony Lineham <anthony.lineham@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>